refactor: mv cmds out of cmd/, seemed nested unecesarrily
