--- src/cmd/mod.rs
@@ -1,3 +1,0 @@
-pub mod build;
-pub mod init;
-pub mod serve;
